Revert "Make src/crypto and src/controller compile with -Wconversion (#2810)" (#2842)
This reverts commit 3ec5a53081fc874641d4635ab43faa1b827e21ab.

The commit included extraneous changes that should not have landed as part of that PR.

+++ src/crypto/tests/PBKDF2_SHA256_test_vectors.h
@@ -34,7 +34,7 @@ struct pbkdf2_test_vector
     const uint8_t * salt;
     size_t slen;
     unsigned int iter;
-    uint32_t key_len;
+    size_t key_len;
     const uint8_t * key;
     unsigned tcId;
     int result;
